refactor: unificar metadata y behavior de providers
